+++ cpu/mem/paged_memory_test.go
@@ -10,7 +10,7 @@ func TestPagedMemory(t *testing.T) {
 
 	memory.WriteStringAt(12, "你好 world.")
 
-	fmt.Printf("%s\n", memory.ReadStringAt(12, uint64(len([]byte("Hello world.")))))
+	fmt.Printf("%s\n", memory.ReadStringAt(12, uint32(len([]byte("Hello world.")))))
 
 	memory.WriteWordAt(1, 12)
